# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Prompt Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Tool {
  slug: string;
  name: string;
  description: string;
  category: string;
  icon?: string;
  modifiers: string[];
  modes: string[];
  seoContent: string;
}

/**
 * Tool-specific generation options (dynamic key/value pairs)
 */
export type PromptGenerateInputOptions = {[key: string]: string};

export interface PromptGenerateInput {
  /** Slug of the tool: logo-prompt, product-photo-prompt, portrait-prompt, youtube-thumbnail-prompt, packaging-prompt */
  toolSlug: string;
  /** Prompt mode: minimal, premium, creative, luxury */
  mode: string;
  /** Tool-specific generation options (dynamic key/value pairs) */
  options: PromptGenerateInputOptions;
  /**
     * Optional session ID for anonymous tracking
     * @nullable
     */
  sessionId?: string | null;
}

export interface QualityScore {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  strengths: string[];
  weaknesses: string[];
  suggestions: string[];
  /** Rating: low, medium, high */
  commercialReadiness: string;
}

export interface PromptFragments {
  /** @nullable */
  introduction?: string | null;
  /** @nullable */
  style?: string | null;
  /** @nullable */
  composition?: string | null;
  /** @nullable */
  lighting?: string | null;
  /** @nullable */
  boosters?: string | null;
  /** @nullable */
  negativePrompt?: string | null;
}

export interface GeneratedPrompt {
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  qualityScore: QualityScore;
  toolSlug: string;
  mode: string;
  fragments?: PromptFragments;
}

export interface HistoryEntry {
  id: number;
  toolSlug: string;
  mode: string;
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  qualityScore?: number | null;
  /**
     * JSON-serialized options
     * @nullable
     */
  options?: string | null;
  /** @nullable */
  sessionId?: string | null;
  createdAt: string;
}

export interface HistoryInput {
  toolSlug: string;
  mode: string;
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  qualityScore?: number | null;
  /** @nullable */
  options?: string | null;
  /** @nullable */
  sessionId?: string | null;
}

export interface FavoriteEntry {
  id: number;
  toolSlug: string;
  mode: string;
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  qualityScore?: number | null;
  /** @nullable */
  sessionId?: string | null;
  createdAt: string;
}

export interface FavoriteInput {
  toolSlug: string;
  mode: string;
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  qualityScore?: number | null;
  /** @nullable */
  sessionId?: string | null;
}

export interface PlatformStats {
  totalPromptsGenerated: number;
  totalFavorites: number;
  totalToolsAvailable: number;
  /** @nullable */
  topToolSlug: string | null;
}

export type ListHistoryParams = {
toolSlug?: string;
limit?: number;
};

export type ListFavoritesParams = {
toolSlug?: string;
};