# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Prompt Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  toolSlug: string;
  mode: string;
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  qualityScore?: number | null;
  /**
     * JSON-serialized options
     * @nullable
     */
  options?: string | null;
  /** @nullable */
  sessionId?: string | null;
  createdAt: string;
}