# lib/api-zod/src/generated/types/promptGenerateInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Prompt Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { PromptGenerateInputOptions } from './promptGenerateInputOptions';

export interface PromptGenerateInput {
  /** Slug of the tool: logo-prompt, product-photo-prompt, portrait-prompt, youtube-thumbnail-prompt, packaging-prompt */
  toolSlug: string;
  /** Prompt mode: minimal, premium, creative, luxury */
  mode: string;
  /** Tool-specific generation options (dynamic key/value pairs) */
  options: PromptGenerateInputOptions;
  /**
     * Optional session ID for anonymous tracking
     * @nullable
     */
  sessionId?: string | null;
}