# lib/api-zod/src/generated/types/qualityScore.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Prompt Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface QualityScore {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  strengths: string[];
  weaknesses: string[];
  suggestions: string[];
  /** Rating: low, medium, high */
  commercialReadiness: string;
}